move reporting error boundary to shared package
