+++ packages/shared/src/errors/ReportingErrorBoundary.tsx
@@ -8,9 +8,9 @@ import {
 } from 'react';
 import { Primitive } from 'relay-runtime/lib/store/RelayStoreTypes';
 
-import { ErrorReportingContext } from '~/shared/contexts/ErrorReportingContext';
-import { CouldNotRenderNftError } from '~/shared/errors/CouldNotRenderNftError';
-import { ErrorWithSentryMetadata } from '~/shared/errors/ErrorWithSentryMetadata';
+import { ErrorReportingContext } from '../contexts/ErrorReportingContext';
+import { CouldNotRenderNftError } from './CouldNotRenderNftError';
+import { ErrorWithSentryMetadata } from './ErrorWithSentryMetadata';
 
 export type ReportingErrorBoundaryFallbackProps = { error: Error };
